i think chaining addition works now, it uses a recursive function

+++ src/generate/generate.c
@@ -30,23 +30,28 @@ void pop(FILE *fptr, const char *reg){
     stack_size--;
 }
 
-void get_var(FILE *fptr, const char *reg, const char *ident){
+void get_var(FILE *fptr, const char *reg, const char *ident, bool use_offset_in, size_t offset_in){
     size_t offset = 0;
-    bool initialized_var = false;
-    for(size_t i = 0; i < stack_vars.size; i++){
-        if(strcmp(stack_vars.data[i].ident, ident) == 0){
-            offset = (stack_size - stack_vars.data[i].stack_loc - 1) * 8;
-            if((int)offset < 0){
-                offset = 0;
+    if(ident != NULL && !use_offset_in){
+        bool initialized_var = false;
+        for(size_t i = 0; i < stack_vars.size; i++){
+            if(strcmp(stack_vars.data[i].ident, ident) == 0){
+                offset = (stack_size - stack_vars.data[i].stack_loc - 1) * 8;
+                if((int)offset < 0){
+                    offset = 0;
+                }
+                initialized_var = true;
+                break;
             }
-            initialized_var = true;
-            break;
+        }
+        if(!initialized_var){
+            fprintf(stderr, "ERROR: Variable, %s, is not initialized\n", ident);
+            system("rm build/out.S");
+            exit(EXIT_FAILURE);
         }
     }
-    if(!initialized_var){
-        fprintf(stderr, "ERROR: Variable, %s, is not initialized\n", ident);
-        system("rm build/out.S");
-        exit(EXIT_FAILURE);
+    else if(use_offset_in){
+        offset = offset_in * 8;
     }
     printf("offset: %li\n", offset);
 
@@ -72,7 +77,7 @@ void expr_add(FILE *fptr, node_stmt *n){
     else if(n->expr->binary.left->term->type == TermType_ident){
         for(size_t i = 0; i < stack_vars.size; i++){
             if(strcmp(n->expr->binary.left->term->ident.val, stack_vars.data[i].ident) == 0){
-                get_var(fptr, "rax", stack_vars.data[i].ident);
+                get_var(fptr, "rax", stack_vars.data[i].ident, false, 0);
                 break;
             }
         }
@@ -84,7 +89,7 @@ void expr_add(FILE *fptr, node_stmt *n){
     else if(n->expr->binary.right->term->type == TermType_ident){
         for(size_t i = 0; i < stack_vars.size; i++){
             if(strcmp(n->expr->binary.right->term->ident.val, stack_vars.data[i].ident) == 0){
-                get_var(fptr, "rdi", stack_vars.data[i].ident);
+                get_var(fptr, "rdi", stack_vars.data[i].ident, false, 0);
                 break;
             }
         }
@@ -98,14 +103,14 @@ void generate_expr(FILE *fptr, node_expr *expr){
         // asm for this
         if (expr->binary.left->type == ExprType_Term){
             if(expr->binary.left->term->type == TermType_ident){
-                get_var(fptr, "rax", expr->binary.left->term->ident.val);
+                get_var(fptr, "rax", expr->binary.left->term->ident.val, false, 0);
             } else{
                 fprintf(fptr, "    mov rax, %s\n", expr->binary.left->term->int_lit.val);
             }
         }
         if (expr->binary.right->type == ExprType_Term){
             if(expr->binary.right->term->type == TermType_ident){
-                get_var(fptr, "rdi", expr->binary.right->term->ident.val);
+                get_var(fptr, "rdi", expr->binary.right->term->ident.val, false, 0);
             } else{
                 fprintf(fptr, "    mov rdi, %s\n", expr->binary.right->term->int_lit.val);
             }
@@ -113,9 +118,10 @@ void generate_expr(FILE *fptr, node_expr *expr){
         else{
             push(fptr, "rax");
             generate_expr(fptr, expr->binary.right);
-            pop(fptr, "rax");
+            // pop(fptr, "rax");
+            get_var(fptr, "rdi", NULL, true, 1);
             fprintf(fptr, "    add rax, rdi\n");
-            get_var(fptr, "rax", )
+
             push(fptr, "rax");
             return;
         }
@@ -126,7 +132,7 @@ void generate_expr(FILE *fptr, node_expr *expr){
     else if(expr->type == ExprType_Term){
         printf("%s\n", expr->term->ident.val);
         if(expr->term->type == TermType_ident){
-            get_var(fptr, "rax", expr->term->ident.val);
+            get_var(fptr, "rax", expr->term->ident.val, false, 0);
             push(fptr, "rax");
         }
         else if(expr->term->type == TermType_int_lit){
@@ -164,7 +170,7 @@ void generate(dynlist_stmt *prog)
                 }
                 else if(n->expr->term->type == TermType_ident){
                     fprintf(fptr, "    mov rax, 60\n");
-                    get_var(fptr, "rdi", n->expr->term->ident.val);
+                    get_var(fptr, "rdi", n->expr->term->ident.val, false, 0);
                     fprintf(fptr, "    syscall\n");
                 }
             }
